AG-17119 - move downloadfile to agstack
