+++ packages/ag-grid-community/src/agStack/utils/download.ts
@@ -1,13 +1,6 @@
-import { _warn } from '../validation/logging';
-
 /** @internal AG_GRID_INTERNAL - Not for public use. Can change / be removed at any time. */
 export function _downloadFile(fileName: string, content: Blob) {
     const win = document.defaultView || window;
-
-    if (!win) {
-        _warn(52);
-        return;
-    }
 
     const element = document.createElement('a');
     const url = win.URL.createObjectURL(content);
